+++ frontend/components/feedback-task/header/filters/LabelsSelector.vue
@@ -5,10 +5,15 @@
     @keyup.up="preselectPreviousOption"
     @keyup.down="preselectNextOption"
   >
-    <MetadataLabelsSelectorSearch
+    <LabelsSelectorSearch
       v-model="searchText"
       :placeholder="$t('filterBy')"
-      :selected-options="metadata.selectedOptions"
+      :selected-options="filter.selectedOptions"
+    />
+    <OptionsSelector
+      v-if="filter.hasOperator"
+      v-model="filter.operator"
+      :options="['and', 'or']"
     />
     <div class="labels-selector__items">
       <BaseCheckbox
@@ -19,12 +24,12 @@
             : null
         "
         v-for="(option, index) in labelsFilteredBySearchText"
-        :key="option.label"
+        :key="option.value"
         :value="option.selected"
         v-model="option.selected"
         @mouseover.native="preSelectionIndex = index"
       >
-        {{ option.label }}
+        {{ option.text ?? option.value }}
       </BaseCheckbox>
     </div>
   </div>
@@ -32,7 +37,7 @@
 <script>
 export default {
   props: {
-    metadata: {
+    filter: {
       type: Object,
       required: true,
     },
@@ -47,7 +52,7 @@ export default {
     searchText() {
       this.preSelectionIndex = 0;
     },
-    metadataOptions: {
+    filterOptions: {
       handler: function () {
         if (this.searchText.length) {
           this.searchText = "";
@@ -58,13 +63,13 @@ export default {
   },
   computed: {
     labelsFilteredBySearchText() {
-      return this.metadata.filterByText(this.searchText);
+      return this.filter.filterByText(this.searchText);
     },
     optionsLength() {
-      return this.metadata.options.length;
+      return this.filter.options.length;
     },
-    metadataOptions() {
-      return this.metadata.options;
+    filterOptions() {
+      return this.filter.options;
     },
   },
   methods: {