+++ frontend/v1/domain/entities/similarity/SimilarityCriteria.test.ts
@@ -4,7 +4,7 @@ describe("SimilarityCriteria ", () => {
   describe("reset should", () => {
     test("set default criteria values", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
 
       criteria.reset();
 
@@ -18,21 +18,21 @@ describe("SimilarityCriteria ", () => {
   describe("isCompleted should", () => {
     test("return true when all criteria are defined", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
 
       expect(criteria.isCompleted).toBe(true);
     });
 
     test("return false when recordId is undefined", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete(undefined, "vectorName", 50, "least");
+      criteria.withValue(undefined, "vectorName", 50, "least");
 
       expect(criteria.isCompleted).toBe(false);
     });
 
     test("return false when vectorName is undefined", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", undefined, 50, "least");
+      criteria.withValue("recordId", undefined, 50, "least");
 
       expect(criteria.isCompleted).toBe(false);
     });
@@ -41,47 +41,61 @@ describe("SimilarityCriteria ", () => {
   describe("isEqual should", () => {
     test("return true when all criteria are equal", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
       const other = new SimilarityCriteria();
-      other.complete("recordId", "vectorName", 50, "least");
+      other.withValue("recordId", "vectorName", 50, "least");
 
       expect(criteria.isEqual(other)).toBe(true);
     });
 
     test("return false when recordId is different", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
       const other = new SimilarityCriteria();
-      other.complete("otherRecordId", "vectorName", 50, "least");
+      other.withValue("otherRecordId", "vectorName", 50, "least");
 
       expect(criteria.isEqual(other)).toBe(false);
     });
 
     test("return false when vectorName is different", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
       const other = new SimilarityCriteria();
-      other.complete("recordId", "othervectorName", 50, "least");
+      other.withValue("recordId", "othervectorName", 50, "least");
 
       expect(criteria.isEqual(other)).toBe(false);
     });
 
     test("return false when limit is different", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
       const other = new SimilarityCriteria();
-      other.complete("recordId", "vectorName", 100, "least");
+      other.withValue("recordId", "vectorName", 100, "least");
 
       expect(criteria.isEqual(other)).toBe(false);
     });
 
     test("return false when order is different", () => {
       const criteria = new SimilarityCriteria();
-      criteria.complete("recordId", "vectorName", 50, "least");
+      criteria.withValue("recordId", "vectorName", 50, "least");
       const other = new SimilarityCriteria();
-      other.complete("recordId", "vectorName", 50, "most");
+      other.withValue("recordId", "vectorName", 50, "most");
 
       expect(criteria.isEqual(other)).toBe(false);
+    });
+  });
+
+  describe("complete", () => {
+    test("should set criteria values from urlParams", () => {
+      const criteria = new SimilarityCriteria();
+      criteria.complete(
+        "record.cc74af18-8b40-4278-a527-83c38868e8f1.vector.mini-lm-sentence-transformers.limit.50.order.most"
+      );
+
+      expect(criteria.recordId).toBe("cc74af18-8b40-4278-a527-83c38868e8f1");
+      expect(criteria.vectorName).toBe("mini-lm-sentence-transformers");
+      expect(criteria.limit).toBe(50);
+      expect(criteria.order).toBe("most");
     });
   });
 });