+++ frontend/v1/infrastructure/services/useRoutes.ts
@@ -7,6 +7,8 @@ type KindOfParam =
   | "_search"
   | "_metadata"
   | "_sort"
+  | "_response"
+  | "_suggestion"
   | "_similarity";
 
 export const ROUTES = {
@@ -60,13 +62,12 @@ export const useRoutes = () => {
     const funcToUse = Object.keys(actualQuery).length ? "push" : "replace";
     let newQuery = {};
 
-    params.forEach(({ key, value, encode }) => {
+    params.forEach(({ key, value }) => {
       if (!value) return;
-      if (encode) value = btoa(value);
 
       newQuery = {
         ...newQuery,
-        [key]: value,
+        [key]: encodeURIComponent(value),
       };
     });
 
@@ -78,19 +79,11 @@ export const useRoutes = () => {
     });
   };
 
-  const getQueryParams = <T>(key: KindOfParam, decode = false): T => {
-    const param = route.value.query[key] as string;
+  const getQueryParams = <T>(key: KindOfParam): T => {
+    const value = route.value.query[key] as string;
+    if (!value) return;
 
-    if (!!param && decode) {
-      try {
-        return atob(param) as T;
-      } catch {
-        // Encrypted param changed manually
-        return undefined;
-      }
-    }
-
-    return param as T;
+    return decodeURIComponent(value) as T;
   };
 
   return {